% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{build_varcor_matrix}
\alias{build_varcor_matrix}
\title{Title}
\usage{
build_varcor_matrix(c1 = NULL, c2 = NULL, rho, phi, psi, error_sd)
}
\arguments{
\item{c1}{}

\item{c2}{}

\item{rho}{}

\item{error_sd}{}

\item{phi_t}{}

\item{phi_s}{}
}
\description{
Title
}
