% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{covDD}
\alias{covDD}
\title{Estimate Covariance Due to Shared Control Individuals between
Difference-in-Differences Estimates}
\usage{
covDD(
  Ndisj_cohort1,
  Ndisj_cohort2,
  Nshared,
  Tpre,
  Tpost,
  Delta,
  rho,
  phi,
  psi,
  sigma_s
)
}
\arguments{
\item{Ndisj_cohort1}{Vector of numbers of unshared individuals in the control
states in cohort 1}

\item{Ndisj_cohort2}{Vector of numbers of unshared individuals in the control
states in cohort 2}

\item{Nshared}{Vector of numbers of shared individuals in the control states}

\item{Tpre}{Number of measurement occasions in the pre-treatment period
(assumed the same across cohorts)}

\item{Tpost}{Number of measurement occasions in the post-treatment period
(assumed the same across cohorts)}

\item{Delta}{Number of measurement occasions between cohort treatment times}

\item{rho}{Vector of within-person correlations for each state in the
analysis. First element is assumed to be for the treated state.}

\item{phi}{Vector of within-period correlations for each state in the
analysis. First element is assumed to be for the treated state.}

\item{psi}{Vector of within-state correlations for each state in the
analysis. First element is assumed to be for the treated state.}

\item{sigma_s}{Vector of standard deviations of the outcome for each state in
the analysis. First element is assumed to be for the treated state in
cohort 1; second element the treated state in cohort 2.}
}
\value{
A scalar estimate of the covariance between the two cohorts' effect
estimates.
}
\description{
Estimate Covariance Due to Shared Control Individuals between
Difference-in-Differences Estimates
}
