% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{expand_vec}
\alias{expand_vec}
\title{Recycle Scalars to Vectors of a Given Length}
\usage{
expand_vec(vec, len)
}
\arguments{
\item{vec}{A vector or scalar}

\item{len}{Desired length}
}
\value{
A vector of length \code{len}.
}
\description{
Checks if a vector has the desired length and throws an error if not, or if
given a scalar argument, recycles it to the desired length.
}
\examples{
expand_vec(1, 5)
expand_vec(c(1, 2), 2)
expand_vec(c(1, 2), 3)
}
