% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cohort.R
\name{generate_cohort}
\alias{generate_cohort}
\title{Generate a Data Skeleton for Staggered Adoption Difference in Differences}
\usage{
generate_cohort(
  nperstate,
  nctrlstates,
  Tobs,
  Tpost,
  txstatename = "tx1",
  cohort_name = txstatename,
  start_time = 0
)
}
\arguments{
\item{nperstate}{number of individuals per state; can be a single integer or
a vector of length \code{ntxstates} + \code{nctrlstates}. If a single integer, it is
assumed that each state has \code{nperstate} individuals.}

\item{nctrlstates}{number of control states}

\item{Tobs}{number of observation periods, assumed to be in months}

\item{Tpost}{Number of observations in the study period for which the
treated state is treated. Must be less than \code{monthly_obs}.}
}
\description{
Create a "long" data frame to serve as a base for simulation studies for
staggered adoption difference in differences methods.
}
