% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{manage_cor_input}
\alias{manage_cor_input}
\title{Convert Vector to \code{data.frame} for Generating Random Effect Variances}
\usage{
manage_cor_input(d, r, rname)
}
\arguments{
\item{d}{A \code{data.frame} containing at least one variable called \code{state}
indicating different units in a stacked DiD analysis (treated and control).}

\item{r}{Numeric vector to add to \code{d}}

\item{rname}{Name of variable to add to \code{d}}
}
\value{
A \code{data.frame}
}
\description{
Given a vector \code{r} of correlations (or other data), add a variable called
\code{rname} to a \code{data.frame} \code{d}.
}
\examples{
x <- manage_cor_input(d = data.frame("state" = c("MD", "MI", "PA")), r = c(0.2, 0.3, 0.4), rname = "rho")
x <- manage_cor_input(d = x, r = c(0.1, 0.2, 0.3), rname = "phi")
x
}
