% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{overlap_durations}
\alias{overlap_durations}
\title{Compute Durations of Time Overlap Intervals}
\usage{
overlap_durations(Tpre, Tpost, Delta)
}
\arguments{
\item{Tpre}{Number of measurement occasions in the pre-treatment period
(assumed common to both cohorts)}

\item{Tpost}{Number of measurement occasions in the post-treatment period
(assumed common to both cohorts)}

\item{Delta}{Number of measurement occasions between cohort 1's treatment
time and cohort 2's}
}
\value{
A named list of overlap interval durations. Names are:
\itemize{
\item \code{pre_dot}: Overlap interval in which the early cohort is in its
pre-treatment period and the later cohort's study period has not yet started.
\item \code{pre_pre}: Both cohorts are in their pre-treatment periods.
\item \code{post_dot}: Early cohort is in its post-treatment period, later cohort's
study period has not yet started.
\item \code{post_pre}: Early cohort is in its post-treatment period, later cohort is
in its pre-treatment period.
\item \code{post_post}: Both cohorts are in their post-treatment periods.
\item \code{dot_pre}: The early cohort's study period has ended, later cohort is in
its pre-treatment period.
\item \code{dot_post}: The early cohort's study period has ended, later cohort is in
its post-treatment period.
}

Note that it is not possible for all overlap interval durations to be
non-zero: at least one overlap interval will not exist for any combination
of positive \code{Tpre}, \code{Tpost}, and \code{Delta}.
}
\description{
When two treated states' study periods overlap in time, the
total duration of the time window covered by the two study periods (i.e.,
from the start of the earlier period to the end of the later period) can be
decomposed into a number of "overlap intervals" based on whether a study
period has started, ended, or is pre- or post-treatment. This function
computes the number of measurement occasions that fall into each of those
possible overlap intervals.
}
\examples{
overlap_durations(Tpre = 3, Tpost = 3, Delta = 2)
}
\seealso{
\link{time_windows}
}
