% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{post_pre_cov}
\alias{post_pre_cov}
\title{Estimate the Correlation Due to Shared Control Individuals between
Post- and Pre-Treatment Means in a Single Control Unit in
Difference-and-Differences}
\usage{
post_pre_cov(
  ndisj_cohort1,
  ndisj_cohort2,
  nshared,
  Tpre,
  Tpost,
  Delta,
  rho,
  phi_t,
  phi_s,
  sigma_s
)
}
\arguments{
\item{ndisj_cohort1}{Number of unshared control individuals in the unit who
contribute to cohort 1}

\item{ndisj_cohort2}{Number of unshared control individuals in the unit who
contribute to cohort 2}

\item{nshared}{Number of shared control individuals in the unit who
contribute to both cohorts}

\item{Tpre}{Number of measurement occasions in the pre-treatment period
(assumed the same across cohorts)}

\item{Tpost}{Number of measurement occasions in the post-treatment period
(assumed the same across cohorts)}

\item{Delta}{Number of measurement occasions between cohort treatment times}

\item{rho}{Within-person correlation for individuals in the unit.}

\item{phi_t}{Within-period correlation for the unit, i.e., correlation
between outcomes in different individuals at the same time.}

\item{phi_s}{Within-state correlation for the unit, i.e., correlation between
outcomes in different individuals at different times.}

\item{sigma_s}{Standard deviation of the outcome in the unit.}
}
\value{
A scalar estimate of the covariance between two pre-treatment means
in a difference-in-differences context.
}
\description{
In stacked difference-in-differences with multip
}
