% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_shared_controls.R
\name{simulate_shared_controls}
\alias{simulate_shared_controls}
\title{Simulate Stacked DiD Cohorts with Shared Control Individuals}
\usage{
simulate_shared_controls(
  Tpre,
  Tpost,
  Delta,
  nctrlstates,
  nperstate_cohort1,
  nperstate_cohort2 = nperstate_cohort1,
  nshared,
  rho = rep(0, 2 + nctrlstates),
  phi_t = rep(0, 2 + nctrlstates),
  phi_s = rep(0, 2 + nctrlstates),
  tx_intshift_cohort1,
  tx_intshift_cohort2 = tx_intshift_cohort1,
  secular_cohort1 = function(x) 0,
  secular_cohort2 = secular_cohort1,
  state_fe_mean = 0,
  error_sd = rep(1, 2 + nctrlstates),
  verbose = FALSE
)
}
\arguments{
\item{Tpre}{Number of measurement occasions in the pre-treatment period}

\item{Tpost}{Number of measurement occasions in the post-treatment period}

\item{Delta}{Number of measurement occasions between cohort study period
start times}

\item{nctrlstates}{Number of control states in each cohort}

\item{nperstate_cohort1}{Number of individuals in each state in cohort 1.
Optionally, a vector of length 2 + nctrlstates that specifies different
numbers for each state in the order (cohort 1 treated, cohort 2 treated,
control states). Coerced to integer.}

\item{nperstate_cohort2}{Number of individuals in each state in cohort 2; see
above.}

\item{nshared}{Number of shared control individuals}

\item{rho}{Exchangeable within-person correlation. If a single number,
assumed constant across all states; else, a vector of length 2 +
nctrlstates.}

\item{phi_t}{Within-period correlation, i.e., correlation between two
observations from different people in the same state at the same time. If a
single number, assumed constant across all states; else, a vector of length
2 + nctrlstates.}

\item{phi_s}{Within-state correlation, i.e., correlation between two
observations from different people in the same state at different times. If
a single number, assumed constant across all states; else, a vector of
length 2 + nctrlstates.}

\item{tx_intshift_cohort1}{Numeric. Intercept shift at time of treatment
\eqn{t^*} for the treated state in cohort 1.}

\item{tx_intshift_cohort2}{Numeric. Intercept shift at time of treatment
\eqn{t^*} for the treated state in cohort 2.}

\item{secular_cohort1}{Function of one argument defining a secular time trend
common to all states in cohort 1. Defaults to 0.}

\item{secular_cohort2}{Function of one argument defining a secular time trend
common to all states in cohort 2. Defaults to 0.}

\item{state_fe_mean}{Mean of state fixed effects. See Details.}

\item{error_sd}{Standard deviation of additive error in data generative
model. See Details.}

\item{verbose}{Logical. Should additional messages be printed?}
}
\description{
Simulate Stacked DiD Cohorts with Shared Control Individuals
}
