% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{time_windows}
\alias{time_windows}
\title{Identify Stages of Overlapping Study Periods in Difference-in-Differences}
\usage{
time_windows(c1Times, c2Times, c1tStar, c2tStar)
}
\arguments{
\item{c1Times}{Vector of times in cohort 1's study period}

\item{c2Times}{Vector of times in cohort 2's study period}

\item{c1tStar}{First treated time in cohort 1. Must be an element of
\code{c1Times}.}

\item{c2tStar}{First treated time in cohort 2. Must be an element of
\code{c2Times}.}
}
\value{
A named list identifying which of \code{c1Times} and \code{c2Times} fall into
each possible pair of pre/post periods.
}
\description{
Identify Stages of Overlapping Study Periods in Difference-in-Differences
}
\examples{
time_windows(c1Times = c(0, 1, 2), c2Times = c(1, 2, 3), c1tStar = 1, c2tStar = 2)
}
\seealso{
\link{overlap_durations}
}
