% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{var_cov_ctrl_pre_post}
\alias{var_cov_ctrl_pre_post}
\title{Compute Covariance between Pre- and Post-Treatment Mean Outcomes Averaged
over All Control States}
\usage{
var_cov_ctrl_pre_post(nctrl, Tpre, Tpost, rho, phi_s, sigma_s)
}
\arguments{
\item{nctrl}{Vector of numbers of individuals in each control state}

\item{Tpre}{Number of measurement occasions in the pre-treatment period}

\item{Tpost}{Number of measurement occasions in the post-treatment period}

\item{rho}{Vector of within-person correlations for each control state}

\item{phi_s}{Vector of between-period correlations for each control state}

\item{sigma_s}{Vector of outcome variances in each control state}
}
\description{
Compute Covariance between Pre- and Post-Treatment Mean Outcomes Averaged
over All Control States
}
