% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{var_cov_tx_pre_post}
\alias{var_cov_tx_pre_post}
\title{Compute Covariance between Pre- and Post-Treatment Mean Outcomes in a Treated State}
\usage{
var_cov_tx_pre_post(ntx, Tpre, Tpost, rho, psi, sigma_tx)
}
\arguments{
\item{ntx}{Number of individuals in the treated unit}

\item{Tpre}{Number of measurement occasions in the pre-treatment period}

\item{Tpost}{Number of measurement occasions in the post-treatment period}

\item{rho}{Within-person correlation}

\item{psi}{Between-period correlation}

\item{sigma_tx}{Variance of outcome in treated unit}
}
\value{
A numeric scalar
}
\description{
Compute Covariance between Pre- and Post-Treatment Mean Outcomes in a Treated State
}
