% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{var_ctrl_period}
\alias{var_ctrl_period}
\title{Compute Variance of Mean Outcome Averaged over Control States in either Pre
or Post-Treatment Period}
\usage{
var_ctrl_period(nctrl, nTimes, rho, phi_t, phi_s, sigma_s)
}
\arguments{
\item{nctrl}{Vector of numbers of individuals in each control state}

\item{nTimes}{Number of measurement occasions in the period (pre or post tx)}

\item{rho}{Vector of within-person correlations for each control state}

\item{phi_t}{Vector of within-period correlations for each control state}

\item{phi_s}{Vector of between-period correlations for each control state}

\item{sigma_s}{Vector of outcome variances in each control state}
}
\description{
Compute Variance of Mean Outcome Averaged over Control States in either Pre
or Post-Treatment Period
}
