% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{var_tx_period}
\alias{var_tx_period}
\title{Compute Variance of Mean Outcome in Treated State in either Pre or
Post-Treatment Period}
\usage{
var_tx_period(ntx, nTimes, rho, phi_t, phi_s, sigma_tx)
}
\arguments{
\item{ntx}{Number of individuals in the treated unit}

\item{nTimes}{Number of measurement occasions in the period (pre or post tx)}

\item{rho}{Within-person correlation}

\item{phi_t}{Within-period correlation}

\item{phi_s}{Between-period correlation}

\item{sigma_tx}{Variance of outcome in treated unit}
}
\value{
A scalar
}
\description{
Compute Variance of Mean Outcome in Treated State in either Pre or
Post-Treatment Period
}
