% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{varDD}
\alias{varDD}
\title{Variance of Single-Treated-State Difference-in-Differences Estimator for the
Average ATT across Pre- and Post-Treatment Measurement Occasions}
\usage{
varDD(ntx, nctrl, Tpre, Tpost, rho, phi_t, phi_s, sigma_s)
}
\arguments{
\item{ntx}{Number of individuals in the single treated state}

\item{nctrl}{Vector of numbers of individuals for each control state}

\item{Tpre}{Number of measurement occasions in the pre-treatment period}

\item{Tpost}{Number of measurement occasions in the post-treatment period}

\item{rho}{Vector of within-person correlations for each state in the
analysis. First element is assumed to be for the treated state.}

\item{phi_t}{Vector of within-period correlations for each state in the
analysis. First element is assumed to be for the treated state.}

\item{phi_s}{Vector of within-state correlations for each state in the
analysis. First element is assumed to be for the treated state.}

\item{sigma_s}{Vector of standard deviations of the outcome for each state in
the analysis. First element is assumed to be for the treated state in
cohort 1; second element the treated state in cohort 2.}
}
\value{
A scalar
}
\description{
Variance of Single-Treated-State Difference-in-Differences Estimator for the
Average ATT across Pre- and Post-Treatment Measurement Occasions
}
