% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{compute_raneff_vars}
\alias{compute_raneff_vars}
\title{Compute}
\usage{
compute_raneff_vars(rho, phi, psi, error_sd, c1 = NULL, c2 = NULL)
}
\arguments{
\item{rho}{}

\item{phi}{}

\item{psi}{}

\item{error_sd}{}

\item{c1}{}

\item{c2}{}
}
\description{
Compute
}
