% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{cormat}
\alias{cormat}
\title{Construct Correlation Matrix}
\usage{
cormat(rho, p, corstr = c("independence", "exchangeable", "ar1"))
}
\arguments{
\item{rho}{Correlation parameter}

\item{p}{Dimension of desired matrix (will be \eqn{p * p})}

\item{corstr}{Correlation structure; one of "independence", "exchangeable",
or "ar1". Partial matching used.}
}
\value{
A \eqn{p * p} matrix of correlations.
}
\description{
Construct Correlation Matrix
}
\examples{
corstr(rho = 0.3, p = 3, corstr = "exchangeable") 
}
