% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_shared_controls.R
\name{simulate_shared_controls}
\alias{simulate_shared_controls}
\title{Simulate Stacked DiD Cohorts with Shared Control Individuals}
\usage{
simulate_shared_controls(
  Tpre,
  Tpost,
  Delta,
  nctrlstates,
  nperstate_cohort1,
  nperstate_cohort2 = nperstate_cohort1,
  nshared,
  rho = rep(0, 2 + nctrlstates),
  phi = rep(0, 2 + nctrlstates),
  psi = rep(0, 2 + nctrlstates),
  tx_intshift_cohort1,
  tx_intshift_cohort2 = tx_intshift_cohort1,
  secular_trend = function(x) 0,
  state_fe_mean = 0,
  error_sd = 1,
  ARdecay = 1
)
}
\arguments{
\item{Tpre}{Number of measurement occasions in the pre-treatment period.
Coerced to integer.}

\item{Tpost}{Number of measurement occasions in the post-treatment period.
Coerced to integer.}

\item{Delta}{Number of measurement occasions between cohort study period
start times. Coerced to integer.}

\item{nctrlstates}{Number of control states; one number common to both
cohorts. Coerced to integer.}

\item{nperstate_cohort1}{Number of individuals in each state in cohort 1.
Recycled to length (\code{nctrlstates} + 1) such that each element specifies
sample size for each state in the order (cohort 1's treated state, control
state 1, control state 2, ...)}

\item{nperstate_cohort2}{Number of individuals in each state in cohort 2; see
\code{nperstate_cohort1} above.}

\item{nshared}{Number of shared control individuals per control state that
contribute to analysis for both cohorts, of the full state sample size
provided in the \code{nperstate} variables. Recycled to length \code{nctrlstates} and
assumed to be given in the order (control state 1, control state 2, ...)}

\item{rho}{Exchangeable within-person correlation. If a single number,
assumed constant across all states; else, a vector of length 2 +
\code{nctrlstates} in the order (cohort 1 treated, cohort 2 treated, control
states).}

\item{phi}{Within-period correlation, i.e., correlation between two
observations from different people in the same state at the same time.
Recycled to length 2 + \code{nctrlstates} in the order (cohort 1 treated, cohort
2 treated, control states).}

\item{psi}{Within-state correlation, i.e., correlation between two
observations from different people in the same state at different times.
Recycled to length 2 + \code{nctrlstates} in the order (cohort 1 treated, cohort
2 treated, control states).}

\item{tx_intshift_cohort1}{Numeric. Intercept shift at time of treatment
\eqn{t^*} for the treated state in cohort 1.}

\item{tx_intshift_cohort2}{Numeric. Intercept shift at time of treatment
\eqn{t^*} for the treated state in cohort 2.}

\item{secular_trend}{Function of one argument (time) defining a secular time
trend common to all states in the analysis. Defaults to 0.}

\item{state_fe_mean}{Mean of state fixed effects. See Details.}

\item{error_sd}{Standard deviation of additive error in data generative
model, assumed common across person-state-time. See Details.}

\item{ARdecay}{Autoregressive parameter for the correlation structure of the
unit-time random effects in the data-generative model. Must be between 0
and 1; 1 indicates no autoregressive decay, 0 implies unit-time random
effects are pairwise independent.}
}
\value{
A list
}
\description{
Generate, summarize, and analyze a two-cohort stacked
difference-in-differences study.
}
\details{
Generates, summarizes, and analyzes a \code{data.frame} for a two-cohort
stacked difference-in-differences analysis with shared control individuals.
The \code{data.frame} contains individual-level data from two treated states and
\code{nctrlstates} control states that are common to both treated states'
analyses.

The generative model for the outcome observed from individual \eqn{i} at
time \eqn{} unit/state \eqn{u} is \deqn{Y_{uit} = \beta_0 + \beta_{1,t} +
  \beta_2 A_{u t} + b_{i} + b_{u} + b_{u t} + \epsilon_{uit},} where
\eqn{\beta_0} corresponds to \code{state_fe_mean}, \eqn{\beta_{1,t} =
  \beta_1(t)} is \verb{secular_trend(}\eqn{t}\verb{)}; \eqn{\beta_2} is
\code{tx_intshift_cohortK}; \eqn{A_{ut}} is an indicator for whether unit
\eqn{u} is treated at time \eqn{t};  \eqn{b_i}, \eqn{b_u}, and \eqn{b_{ut}}
are random effects for individual, unit, and unit-time, respectively; and
\eqn{\epsilon_{uit}} is residual error with standard deviation \code{error_sd}.
In the above, \code{K} is either 1 or 2 depending on the cohort for which data
is being generated. Note that state fixed effects are \emph{not}
user-specified: they are defined as \eqn{\beta_0 + b_u}, with \eqn{b_u \sim
  N(0, \sigma^2_{bu})}. Also, note that the treatment effect is \emph{constant}
over time.

The analytic
}
\examples{
simulate_shared_controls(
    Tpre = 3,
    Tpost = 2, 
    Delta = 1,
    nctrlstates = 4,
    nperstate_cohort1 = 100,
    nperstate_cohort2 = 150,
    nshared = 50,
    tx_intshift_cohort1 = 1,
    tx_intshift_cohort2 = 1.5
)
}
