% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDD_helpers.R
\name{var_ctrl_period}
\alias{var_ctrl_period}
\title{Compute Variance of Mean Outcome Averaged over Control States in either Pre-
or Post-Treatment Period}
\usage{
var_ctrl_period(nctrl, nTimes, rho, phi, psi, sigma_s)
}
\arguments{
\item{nctrl}{Vector of numbers of individuals in each control state}

\item{nTimes}{Number of measurement occasions in the period (pre or post tx)}

\item{rho}{Vector of within-person correlations for each control state}

\item{phi}{Vector of within-period correlations for each control state}

\item{psi}{Vector of between-period correlations for each control state}

\item{sigma_s}{Vector of outcome variances in each control state}
}
\value{
A scalar, the variance of the mean outcome in control states over
either the pre- or post-treatment period
}
\description{
Compute Variance of Mean Outcome Averaged over Control States in either Pre-
or Post-Treatment Period
}
